Relatório html, modificado o tamanho da janela do navedador e captura de tela dos testes.
