--- src/test/resources/cucumber.properties
@@ -1,2 +1,0 @@
-cucumber.publish.quiet=true
-cucumber.snippet-type=camelcase
